# storage/api-docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Garage - Authentification JWT",
    "description": "API d'authentification pour les professionnels (JWT)",
    "version": "1.0.0",
    "contact": {
      "name": "TOOAUTO Garage"
    }
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Token JWT (format: Bearer {token})"
    }
  },
  "security": [],
...
  "paths": {
...
    "/vidanges/last-by-matricule": {
      "get": {
        "tags": ["Vidanges"],
        "summary": "Dernière vidange par matricule",
        "description": "Récupère la dernière vidange (alert) d'un véhicule par son matricule",
        "operationId": "getLastVidangeByMatricule",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "matricule", "in": "query", "required": true, "type": "string", "description": "Matricule du véhicule" }
        ],
        "responses": {
          "200": { "description": "Dernière vidange trouvée", "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" }, "vidange": { "$ref": "#/definitions/Alert" } } } },
          "401": { "description": "Non authentifié" },
          "404": { "description": "Établissement, véhicule ou vidange non trouvé" },
          "422": { "description": "Validation échouée (matricule requis)" }
        }
      }
    },
    "/vidanges/vehicule-by-matricule": {
      "get": {
        "tags": ["Vidanges"],
        "summary": "Véhicule par matricule",
        "description": "Récupère un véhicule par son matricule",
        "operationId": "getVehiculeByMatricule",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "matricule", "in": "query", "required": true, "type": "string", "description": "Matricule du véhicule" }
        ],
        "responses": {
          "200": { "description": "Véhicule trouvé", "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" }, "vehicule": { "type": "object" } } } },
          "401": { "description": "Non authentifié" },
          "404": { "description": "Établissement ou véhicule non trouvé" },
          "422": { "description": "Validation échouée (matricule requis)" },
          "500": { "description": "Erreur serveur" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Alert": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "etablissement_id": { "type": "integer" },
        "vehicule_id": { "type": "integer" },
        "type_alert_id": { "type": "integer" },
        "user_id": { "type": "integer" },
        "message": { "type": "string" },
        "statut": { "type": "integer" },
        "date_debut": { "type": "string", "format": "date" },
        "date_fin": { "type": "string", "format": "date" },
        "kilometrage": { "type": "string" },
        "autres": { "type": "string" },
        "created_at": { "type": "string", "format": "date-time" },
        "updated_at": { "type": "string", "format": "date-time" },
        "vehicule": { "type": "object" },
        "typeAlert": { "type": "object" }
      }
    },
...
  }
}